Split up integration test
